Qualify "coplanar" statement in documentation
The code (specifically line https://github.com/evanw/csg.js/blob/a8512afbac3cf503195870f7ef11c0a32f36c6d4/csg.js#L439) uses the word "coplanar" in two ways: coplanarity of polygons with the same direction normal and and coplanarity of polygons with reversed normal directions. The documentation, on the other hand, uses it in one specific way: coplanar polygons with the same normal direction but doesn't quality this word. `clipTo` returns faces with the same normal direction, but removes coplanar polygons with a reversed normal (unless the BSP is degenerate). I think this statement is key to understanding the algorithm correctly, so I think this deserves to be added.